[gh-417] Disable access to Content->Files tab and Media settings
We don't want anyone touching the settings, they can be managed if need be
in the hook_update process.  The Files tab showing Media just will cause
confusion for now.  The point of this module is to enable a simple
WYSIWYG button to add videos, as the Media project matures utilizing more
of it can be considered.